@@ -1,7 +1,7 @@
 class ApplicationController < ActionController::Base
- 
+
   protect_from_forgery with: :exception
-  
+
   include SessionsHelper
 
   private
@@ -11,7 +11,5 @@ class ApplicationController < ActionController::Base
       redirect_to login_url
     end
   end
-  
-  
   
 end
