@@ -1,13 +1,12 @@
 class MemosController < ApplicationController
   before_action :require_user_logged_in
   before_action :correct_user, only: [:destroy]
-  
-  
-  def index 
+
+
+  def index
       @memos = Memo.all.order('created_at DESC').page(params[:page])
-      
-  end  
-  
+  end
+
   def new
       @memo = current_user.memos.build  # form_for 用
       @memos = current_user.memos.order('created_at DESC').page(params[:page])
@@ -31,11 +30,11 @@ class MemosController < ApplicationController
     flash[:success] = 'メッセージを削除しました。'
     redirect_back(fallback_location: root_path)
   end
-  
+
   def edit
     @memo = Memo.find(params[:id])
-  end 
-  
+  end
+
   def update
     @memo = Memo.find(params[:id])
 
@@ -47,13 +46,13 @@ class MemosController < ApplicationController
       render :edit
     end
   end
-  
+
    private
 
   def memo_params
     params.require(:memo).permit(:trick, :content)
   end
-  
+
   def correct_user
     @memo = current_user.memos.find_by(id: params[:id])
     unless @memo
