@@ -1,11 +1,10 @@
 class RelationshipsController < ApplicationController
   def new
     
-    
   end
 
   def create
-    
+
   end
 
   def destroy
