@@ -11,9 +11,8 @@ class TricksController < ApplicationController
   def create
     genre = Genre.find(trick_params[:genre_id])
     @trick = genre.tricks.new(trick_params[:name])
-    
-      
-    if @trick.save 
+
+    if @trick.save
       flash[:success] = 'トリックを登録しました。'
       redirect_to root_url
     else
@@ -24,11 +23,11 @@ class TricksController < ApplicationController
 
   def destroy
   end
-  
+
   private
-  
+
   def trick_params
     params.require(:trick).permit(:name, :name, { :genre_id => [] })
   end
-  
+
 end
