@@ -1,6 +1,6 @@
 class UsersController < ApplicationController
   before_action :require_user_logged_in, only: [:show]
-  
+
   def show
     @user = User.find(params[:id])
     @memos = @user.memos.order('created_at DESC').page(params[:page])
@@ -21,13 +21,11 @@ class UsersController < ApplicationController
       render :new
     end
   end
-  
+
   def library
     @user = User.find(params[:id])
     @liked = @user.liked.page(params[:page])
-    
-    
-  end  
+  end
 
   private
 
