@@ -1,5 +1,5 @@
 module UsersHelper
-    def gravatar_url(user, options = { size: 80 })
+  def gravatar_url(user, options = { size: 80 })
     gravatar_id = Digest::MD5::hexdigest(user.email.downcase)
     size = options[:size]
     "https://secure.gravatar.com/avatar/#{gravatar_id}?s=#{size}&d=mp"
